% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a FARS query into a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Relative filepath to the current working directory. This must be an existing comma separated values file (*.csv) or it will throw an error.}
}
\value{
A tibble from the inputted *.csv file.
}
\description{
fars_read() takes in a *.csv file from the NHTSA's FARS #' archives  and returns a tibble.
}
\details{
This function takes in a *.csv file and using the readr package will load the file as a tibble then will return it as a tbl_df (subclass of a data.frame) through dplyr. This function is ideally to be used with the FARS (Fatality Analysis Reporting System) archives for accident data in *.csv format. If no file exists an error will occur before attempting to read the file.
}
\examples{
# this will stop
#returns data as tbl_df
  \dontrun{
     fars_read("this_file_does_not_exist.csv")
  }

#setwd(system.file("data-raw", package = "wk4package"))
#getwd()
  \dontrun{
     fars_read("accident_2013.csv.bz2")
  }

}
\seealso{
\url{https://www.nhtsa.gov/file-downloads?p=nhtsa/downloads/FARS/} for other .csv files from FARS (Fatality Analysis Reporting System)
to process. Explore the NHTSA's website for more information on the FARS program.

Other FARS functions: 
\code{\link{fars_map_state}()},
\code{\link{fars_read_years}()}
}
\concept{FARS functions}
\keyword{internal}
