% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Search and print out a dataset by month and year of accident.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character/numerical vector containing the years that would like to be query.}
}
\value{
A list of tibbles with the column arrangement of month/year. Years type is dependent on the parameter's type.
}
\description{
Provide a vector of years and receive all accident data for those given years from FARS accident archival data.
}
\details{
Provide a list/vector of years you would like to query in the form of XXXX in numerical, integer, or character form. This will return a list
of each years accident info containing the month and year the accident occured.
}
\examples{
 #read in one year, will return a tibble containing the month as an integer
#setwd(system.file("data-raw", package = "wk4package"))
 \dontrun{fars_read_years(2015)}
 # read in multiple years, will return a list of tibbles 
 # containing the month as a double and year as a <dbl>.
 \dontrun{fars_read_years(c(2013,2014))}
 # You can mix and match object types
 \dontrun{fars_read_years(list("2013",2014))}

}
\seealso{
Other FARS functions: 
\code{\link{fars_map_state}()},
\code{\link{fars_read}()}
}
\concept{FARS functions}
