% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS accident data by month and year;}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Character/numeric vector containing the years to summarize.}
}
\value{
A tibble containing a summary with the years as columns and month as rows. The values are the number of accidents for that year's month.
}
\description{
Summarize the Fatality Analysis Reporting System data, counting the number of accidents per month over the stated set of years.
}
\details{
Do note that the following packages are needed: dplyr, tidyr, readr, magrittr, and purrr.
}
\examples{
#setwd(system.file("data-raw", package = "wk4package"))
# Returns a table with the number of accidents for every month in 2013 and 2014.
\dontrun{fars_summarize_years(c(2013, 2014))}
# Character objects work too
\dontrun{fars_summarize_years(c("2013", "2014"))}
}
