% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate filename for FARS data.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A "character" or "numeric" input that can be interpreted as an integer by the \code{as.integer()} function.}
}
\value{
A "character" object where the number
}
\description{
Input a year that you want to receive the FARS filename of.
}
\details{
Input a numeric or character object representing an integer which will then be coerced into an integer object. This integer is finally formatted into a character object of the form "accident_1234.csv.bz2" where "1234" is placeholder for the integer.
}
\examples{
#setwd(system.file("data-raw", package = "wk4package"))
#returns "accident_2009.csv.bz2"
\dontrun{make_filename(2009)}
#returns "accident_2012.csv.bz2"
\dontrun{make_filename("2012")}
}
